Handler: add signals to update state
Sometimes, you need to update the state of a stateful Handler,
without using a Mississippi message.
Introduce the concept of 'signal': an information that is
directly forwarded to the handler process, without
being tracked. The signal is still identified by the
sharding key of the handler and can be any term.
Update the `Mississippi.Consumer.DataUpdater.Handler` behaviour
accordingly. Calling for a signal to be handled is a blocking
operation.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>