Consumer: new MessageTracker
A MessageTracker process is now an Orleans-like grain, responsible
of keeping the order of messages for a given sharding key.
It holds a reference to the channel on which to ack/reject
successfully processed messages, and maintains the invariant
that only the first message in its process queue is being
processed at any given time.

If a DataConsumer fails, the MessageTracker removes from its own queue
all messages that held references to the DataConsumer's channel. Those
messages are automatically requeued from the AMQP library, thus
will be handled again to the MessageTracker, together with a reference
to a new, up, channel.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>