Consumer: new DataUpdater
The DataUpdater process takes care of handling messages and signals only,
for a given sharding key. Ordering is guaranteed by the MessageTracker.

If a DataUpdater were to crash, the MessageTracker will request another
for handling the message. It is thus possible that a message is repeated
twice, but not that it is lost.